feat: 권한에 맞게 접근 경로 설정
