+++ src/main/java/nextstep/infrastructure/annotation/Login.java
@@ -1,4 +1,4 @@
-package nextstep.config;
+package nextstep.infrastructure.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
