+++ src/main/java/nextstep/infrastructure/auth/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package nextstep.infrastructure;
+package nextstep.infrastructure.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.util.Strings;
