+++ src/test/java/nextstep/member/MemberE2ETest.java
@@ -3,7 +3,7 @@ package nextstep.member;
 import io.restassured.RestAssured;
 import nextstep.auth.TokenRequest;
 import nextstep.auth.TokenResponse;
-import nextstep.auth.role.Role;
+import nextstep.infrastructure.role.Role;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
