+++ src/test/java/nextstep/reservation/ReservationE2ETest.java
@@ -3,10 +3,9 @@ package nextstep.reservation;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.auth.AuthService;
 import nextstep.auth.TokenRequest;
 import nextstep.auth.TokenResponse;
-import nextstep.auth.role.Role;
+import nextstep.infrastructure.role.Role;
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import nextstep.member.MemberRequest;
